local_requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Dance==7.1.0
WTForms==3.2.1
email-validator==2.2.0
Werkzeug==3.1.3
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
PyJWT==2.10.1
oauthlib==3.2.2
PyMuPDF==1.26.0
gunicorn==23.0.0
python-dotenv==1.0.0